+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mapwise/core
 
+## 1.0.2
+
+### Patch Changes
+
+- Relax React peer dependencies to support ^18.0.0 || ^19.0.0
+
 ## 1.0.1
 
 ### Patch Changes