+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mapwise/plugins
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @mapwise/core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes